+++ lang/it/menu.php
@@ -52,6 +52,7 @@ return [
     'domande'                 => 'Domande',
     'segnalazioni'            => 'Segnalazioni',
     'tipi_patente'            => 'Tipi di patente',
+    'moduli_guida'            => 'Moduli guida pratica',
     'quizzes'                 => 'Quizzes',
     'i_miei_studenti'         => 'I miei studenti',
     'gestione_istruttori'     => 'Gestione istruttori',